refactor: move the package
- Rank enum to enums package
