+++ src/main/kotlin/lotto/model/Game.kt
@@ -1,7 +1,7 @@
 package lotto.model
 
 import lotto.Lotto
-import lotto.Rank
+import lotto.enums.Rank
 import java.util.*
 
 class Game(private val purchaseAmount: Int, private val lottoTicket: List<Lotto>, private val winningLotto: WinningLotto) {
