+++ src/test/kotlin/lotto/RankTest.kt
@@ -1,5 +1,6 @@
 package lotto
 
+import lotto.enums.Rank
 import org.junit.jupiter.api.Assertions.*
 import org.junit.jupiter.api.Test
